Chirper: Add chirp component to the index page
Following step 3 in the bootcamp:
https://bootcamp.laravel.com/inertia/showing-chirps

We can add new chirps but didn't have a way of viewing the chirps
we've created.

This adds the "chirp" component to view the chirp values and the
dayjs NPM package for relative dates.

The chirp component is shown on the chirp index page, filled will
all the latest chirps.