Chirper: Add "StoreChirpRequest" for validation logic
When we create a new "chirp" we verify that the text is between some
parameters.

The logic to create a new and edit a chirp should be the same, but
we didn't have any actual code link to keep it that way.

This add teh "StoreChirpRequest" which has the common validate logic
we want whenever we modify a chirp.

Updating the logic to handle it was basically just changing the
parameter type.